refactor service

+++ src/main/java/com/miro/api/widgets/testtask/services/WidgetService.java
@@ -5,9 +5,9 @@ import com.miro.api.widgets.testtask.entities.WidgetEntity;
 import java.util.List;
 
 public interface WidgetService {
-    WidgetEntity createWidget(Integer xCoordinate, Integer yCoordinate, Integer z, Integer height, Integer width);
+    WidgetEntity createWidget(int xCoordinate, int yCoordinate, int z, int height, int width);
     WidgetEntity getWidgetById(String id);
-    WidgetEntity updateWidgetById(String id, Integer xCoordinate, Integer yCoordinate, Integer z, Integer height, Integer width);
+    WidgetEntity updateWidgetById(String id, int xCoordinate, int yCoordinate, int z, int height, long width);
     Boolean deleteWidgetById(String id);
     List<WidgetEntity> getAllWidgets();
 }
